refactor the account aside, so we can implement it as well for other pages
