+++ packages/apps/explorer/src/pages/[networkSlug]/transaction/[requestKey].tsx
@@ -1,6 +1,8 @@
 import type { TransactionRequestKeyQuery } from '@/__generated__/sdk';
 import { useTransactionRequestKeyQuery } from '@/__generated__/sdk';
+import { LayoutAside } from '@/components/Layout/components/LayoutAside';
 import { LayoutBody } from '@/components/Layout/components/LayoutBody';
+import { LayoutCard } from '@/components/Layout/components/LayoutCard';
 import { LayoutHeader } from '@/components/Layout/components/LayoutHeader';
 import { Layout } from '@/components/Layout/Layout';
 import { loadingTransactionData } from '@/components/LoadingSkeleton/loadingData/loadingDataTransactionRequestKeyQuery';
@@ -66,12 +68,17 @@ const Transaction: React.FC = () => {
   }, [loading, data, error]);
 
   return (
-    <Layout layout="full">
+    <Layout>
       {innerData && innerData.transaction ? (
         <>
           <LayoutHeader>
             <ValueLoader isLoading={isLoading}>Transaction Details</ValueLoader>
           </LayoutHeader>
+
+          <LayoutAside>
+            <LayoutCard>sdsdf</LayoutCard>
+            <LayoutCard>sdsdf</LayoutCard>
+          </LayoutAside>
 
           <LayoutBody>
             <Tabs isCompact isContained>
